Remove the types file, not enough types to justify
