+++ components/src/core/InfoListItem/InfoListItem.tsx
@@ -1,11 +1,36 @@
 import React from 'react';
 import clsx from 'clsx';
 import { separate, withKeys } from '../utilities';
-import {InfoListItemClasses, InfoListItemProps} from "./InfoListItem.types";
-import { useStyles } from './InfoListItem.styles';
+import {InfoListItemClasses, useStyles} from './InfoListItem.styles';
 
 import { Avatar, Divider, ListItem, ListItemAvatar, ListItemIcon, ListItemText, Typography } from '@material-ui/core';
 import Chevron from '@material-ui/icons/ChevronRight';
+import {CSSProperties} from "@material-ui/core/styles/withStyles";
+
+export type DividerType = 'full' | 'partial';
+export type InfoListItemProps = {
+    avatar?: boolean;
+    backgroundColor?: string;
+    chevron?: boolean;
+    classes?: InfoListItemClasses;
+    dense?: boolean;
+    divider?: DividerType;
+    fontColor?: string;
+    hidePadding?: boolean;
+    icon?: JSX.Element;
+    iconColor?: string;
+    leftComponent?: JSX.Element;
+    onClick?: Function;
+    rightComponent?: JSX.Element;
+    ripple?: boolean;
+    statusColor?: string;
+    style?: CSSProperties;
+    subtitle?: string | Array<string | JSX.Element>;
+    subtitleSeparator?: string;
+    title: string;
+    wrapSubtitle?: boolean;
+    wrapTitle?: boolean;
+};
 
 const MAX_SUBTITLE_ELEMENTS = 6;
